refactor(ui): split spinner and editor into own files
internal/ui/ui.go was 1813 lines covering everything from ANSI colors
to fzf wrappers to TUI prompts to spinners to $EDITOR launching. Two
sections come out cleanly without disturbing the rest:

  - spinner.go (122 lines): SpinnerDelay, Spinner, DelayedSpinner,
    NewSpinner / NewDelayedSpinner, WithSpinner. The spinner is the
    only thing in the package that touches sync and time, so those
    imports leave ui.go entirely.

  - editor.go (78 lines): EditWithEditor — the YesMode-aware,
    TTY-aware $EDITOR launcher. Self-contained.

ui.go drops 184 lines (1813 → 1629) and two unused imports. No
behavior change — spinner.go and editor.go remain in package ui, so
all existing call sites are untouched. The package's full test suite
passes unchanged.

This pairs with the Phase 5 hook in internal/git: now that the spinner
is hidden behind git.ProgressStart and only consumed by cmd/ezs, it
makes sense for it to also be its own file in ui.
