+++ internal/ui/ui.go
@@ -9,8 +9,6 @@ import (
 	"os/exec"
 	"path/filepath"
 	"strings"
-	"sync"
-	"time"
 
 	"github.com/KulkarniKaustubh/ezstack/v4/internal/config"
 	"github.com/chzyer/readline"
@@ -1628,186 +1626,4 @@ func (t *TerminalBackend) SelectOptionWithBack(options []string, prompt string) 
 	}
 
 	return idx - 1, nil
-}
-
-// SpinnerDelay is the delay before showing a spinner (only show for slow operations)
-const SpinnerDelay = 1500 * time.Millisecond
-
-// Spinner represents a simple loading spinner
-type Spinner struct {
-	message  string
-	stop     chan bool
-	wg       sync.WaitGroup
-	started  bool
-	mu       sync.Mutex
-	stopOnce sync.Once
-}
-
-// NewSpinner creates a new spinner with the given message
-func NewSpinner(message string) *Spinner {
-	return &Spinner{
-		message: message,
-		stop:    make(chan bool),
-	}
-}
-
-// Start begins the spinner animation
-func (s *Spinner) Start() {
-	s.wg.Add(1)
-	go func() {
-		defer s.wg.Done()
-		frames := []string{"⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧", "⠇", "⠏"}
-		i := 0
-		for {
-			select {
-			case <-s.stop:
-				// Clear the spinner line if it was shown
-				s.mu.Lock()
-				if s.started {
-					fmt.Fprintf(os.Stderr, "\r\033[K")
-				}
-				s.mu.Unlock()
-				return
-			default:
-				s.mu.Lock()
-				s.started = true
-				s.mu.Unlock()
-				fmt.Fprintf(os.Stderr, "\r%s%s%s %s", Cyan, frames[i], Reset, s.message)
-				i = (i + 1) % len(frames)
-				time.Sleep(80 * time.Millisecond)
-			}
-		}
-	}()
-}
-
-// Stop stops the spinner
-func (s *Spinner) Stop() {
-	s.stopOnce.Do(func() {
-		close(s.stop)
-	})
-	s.wg.Wait()
-}
-
-// DelayedSpinner represents a spinner that only shows after a delay
-type DelayedSpinner struct {
-	message string
-	delay   time.Duration
-	spinner *Spinner
-	timer   *time.Timer
-	mu      sync.Mutex
-	stopped bool
-}
-
-// NewDelayedSpinner creates a spinner that only shows after the configured delay
-func NewDelayedSpinner(message string) *DelayedSpinner {
-	return &DelayedSpinner{
-		message: message,
-		delay:   SpinnerDelay,
-	}
-}
-
-// Start begins the delayed spinner (will only show if not stopped before delay)
-func (ds *DelayedSpinner) Start() {
-	ds.mu.Lock()
-	defer ds.mu.Unlock()
-
-	ds.timer = time.AfterFunc(ds.delay, func() {
-		ds.mu.Lock()
-		defer ds.mu.Unlock()
-		if !ds.stopped {
-			ds.spinner = NewSpinner(ds.message)
-			ds.spinner.Start()
-		}
-	})
-}
-
-// Stop stops the delayed spinner
-func (ds *DelayedSpinner) Stop() {
-	ds.mu.Lock()
-	defer ds.mu.Unlock()
-
-	ds.stopped = true
-	if ds.timer != nil {
-		ds.timer.Stop()
-	}
-	if ds.spinner != nil {
-		ds.spinner.Stop()
-	}
-}
-
-// WithSpinner runs a function with a delayed spinner
-// The spinner only shows if the function takes longer than SpinnerDelay
-func WithSpinner(message string, fn func() error) error {
-	spinner := NewDelayedSpinner(message)
-	spinner.Start()
-	defer spinner.Stop()
-	return fn()
-}
-
-// EditWithEditor opens the user's preferred editor with the initial content
-// and returns the edited content. If the user saves and exits, the content is returned.
-// If the user aborts (empty file or error), an error is returned.
-// The editor is determined by $EDITOR, $VISUAL, or falls back to "vim".
-//
-// Under YesMode there's no controlling terminal for $EDITOR to draw on, so
-// the editor would hang indefinitely. Skip it and return the initial
-// content unchanged — MCP / -y callers that want to edit should pass the
-// content via the relevant flag instead.
-func EditWithEditor(initialContent, fileExtension string) (string, error) {
-	if YesMode {
-		fmt.Fprintf(os.Stderr, "%sezs: skipping interactive editor in non-interactive mode (using initial content)%s\n", Gray, Reset)
-		return strings.TrimSpace(initialContent), nil
-	}
-	// Belt-and-braces: even when YesMode is off, an MCP-style invocation
-	// has no TTY for $EDITOR to draw on. Detect that and short-circuit
-	// rather than launch vim against /dev/null and hang. The TTY check
-	// covers stdin/stdout — most editors require both. We deliberately
-	// require BOTH to be terminals so a user redirecting only stdout
-	// (e.g. `ezs ... > file`) is not misclassified.
-	if !term.IsTerminal(int(os.Stdin.Fd())) || !term.IsTerminal(int(os.Stdout.Fd())) {
-		fmt.Fprintf(os.Stderr, "%sezs: no terminal attached, skipping interactive editor%s\n", Gray, Reset)
-		return strings.TrimSpace(initialContent), nil
-	}
-	// Get the editor from environment
-	editor := os.Getenv("EDITOR")
-	if editor == "" {
-		editor = os.Getenv("VISUAL")
-	}
-	if editor == "" {
-		editor = "vim"
-	}
-
-	// Create a temporary file with the specified extension
-	tmpFile, err := os.CreateTemp("", "ezstack-*"+fileExtension)
-	if err != nil {
-		return "", fmt.Errorf("failed to create temp file: %w", err)
-	}
-	tmpPath := tmpFile.Name()
-	defer os.Remove(tmpPath)
-
-	// Write initial content to the file
-	if _, err := tmpFile.WriteString(initialContent); err != nil {
-		tmpFile.Close()
-		return "", fmt.Errorf("failed to write to temp file: %w", err)
-	}
-	tmpFile.Close()
-
-	// Open the editor
-	cmd := exec.Command(editor, tmpPath)
-	cmd.Stdin = os.Stdin
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-
-	if err := cmd.Run(); err != nil {
-		return "", fmt.Errorf("editor exited with error: %w", err)
-	}
-
-	// Read the edited content
-	content, err := os.ReadFile(tmpPath)
-	if err != nil {
-		return "", fmt.Errorf("failed to read edited file: %w", err)
-	}
-
-	result := strings.TrimSpace(string(content))
-	return result, nil
 }
